requirements.txt
asgiref==3.6.0
boto3==1.26.124
botocore==1.29.124
Django==4.2.1
django-cors-headers==3.14.0
jmespath==0.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
